# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "StackClass API Reference",
    "description": "Backend API and services for StackClass",
    "license": {
      "name": ""
    },
    "version": "0.19.0"
  },
  "paths": {
    "/v1/courses": {
      "get": {
        "tags": [
          "Course"
        ],
        "summary": "Find all courses.",
        "operationId": "find-all-courses",
        "responses": {
          "200": {
            "description": "Courses retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CourseResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Course"
        ],
        "summary": "Create a course.",
        "operationId": "create-course",
        "requestBody": {
          "description": "Create course request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCourseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Course created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseResponse"
                }
              }
            }
          },
          "500": {
            "description": "Failed to create course"
          }
        }
      }
    },
    "/v1/courses/{slug}": {
      "get": {
        "tags": [
          "Course"
        ],
        "summary": "Get a course.",
        "operationId": "get-course-detail",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "description": "The slug of course",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Course retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseDetailResponse"
                }
              }
            }
          },
          "404": {
            "description": "Course not found"
          },
          "500": {
            "description": "Failed to get course"
          }
        }
      },
      "delete": {
        "tags": [
          "Course"
        ],
        "summary": "Delete a course.",
        "operationId": "delete-course",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "description": "The slug of course",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Course deleted successfully"
          },
          "404": {
            "description": "Course not found"
          },
          "500": {
            "description": "Failed to delete course"
          }
        }
      },
      "patch": {
        "tags": [
          "Course"
        ],
        "summary": "Update course from git repository",
        "operationId": "update-course",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "description": "The slug of course",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Course updated successfully"
          },
          "404": {
            "description": "Course not found"
          },
          "500": {
            "description": "Failed to update course"
          }
        }
      }
    },
    "/v1/courses/{slug}/extensions": {
      "get": {
        "tags": [
          "Extension"
        ],
        "summary": "Find all extensions for a course.",
        "operationId": "find-all-extensions",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "description": "The slug of course",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Extensions retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ExtensionResponse"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Course not found"
          },
          "500": {
            "description": "Failed to get course"
          }
        }
      }
    },
...
    "/v1/user/courses": {
      "get": {
        "tags": [
          "User",
          "Course"
        ],
        "summary": "Find all courses for the current user.",
        "operationId": "find-user-courses",
        "responses": {
          "200": {
            "description": "User courses retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserCourseResponse"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWTBearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "User",
          "Course"
        ],
        "summary": "Enroll the current user in a course.",
        "operationId": "enroll-user-in-course",
        "requestBody": {
          "description": "Enroll user in course request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserCourseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "User enrolled in course successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserCourseResponse"
                }
              }
            }
          },
          "404": {
            "description": "Course not found"
          },
          "500": {
            "description": "Failed to enroll user in course"
          }
        },
        "security": [
          {
            "JWTBearerAuth": []
          }
        ]
      }
    },
    "/v1/user/courses/{slug}": {
      "get": {
        "tags": [
          "User",
          "Course"
        ],
        "summary": "Get the course detail for the current user.",
        "operationId": "get-user-course-detail",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "description": "The slug of course",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Course retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserCourseResponse"
                }
              }
            }
          },
          "404": {
            "description": "Course not found"
          },
          "500": {
            "description": "Failed to get course"
          }
        },
        "security": [
          {
            "JWTBearerAuth": []
          }
        ]
      },
      "patch": {
        "tags": [
          "User",
          "Course"
        ],
        "summary": "Update this course for the current user.",
        "operationId": "update-user-course",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "description": "The slug of course",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Update this course for the current user",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserCourseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "User course updated successfully"
          },
          "404": {
            "description": "Course not found"
          },
          "500": {
            "description": "Failed to enroll user in course"
          }
        },
        "security": [
          {
            "JWTBearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "CourseDetailResponse": {
        "type": "object",
        "required": [
          "slug",
          "name",
          "short_name",
          "release_status",
          "description",
          "summary",
          "logo",
          "stage_count",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Creation timestamp"
          },
          "description": {
            "type": "string",
            "description": "Detailed description"
          },
          "logo": {
            "type": "string",
            "description": "URL or path to the course logo"
          },
          "name": {
            "type": "string"
          },
          "release_status": {
            "type": "string",
            "description": "Release status (alpha/beta/live)"
          },
          "short_name": {
            "type": "string"
          },
          "slug": {
            "type": "string",
            "description": "Unique human-readable identifier"
          },
          "stage_count": {
            "type": "integer",
            "format": "int32",
            "description": "Number of stages in the course"
          },
          "summary": {
            "type": "string",
            "description": "Brief summary"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "Last update timestamp"
          }
        }
      },
      "CourseResponse": {
        "type": "object",
        "required": [
          "slug",
          "name",
          "short_name",
          "release_status",
          "summary",
          "logo",
          "stage_count",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Creation timestamp"
          },
          "logo": {
            "type": "string",
            "description": "URL or path to the course logo"
          },
          "name": {
            "type": "string"
          },
          "release_status": {
            "type": "string",
            "description": "Release status (alpha/beta/live)"
          },
          "short_name": {
            "type": "string"
          },
          "slug": {
            "type": "string",
            "description": "Unique human-readable identifier"
          },
          "stage_count": {
            "type": "integer",
            "format": "int32",
            "description": "Number of stages in the course"
          },
          "summary": {
            "type": "string",
            "description": "Brief summary"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "Last update timestamp"
          }
        }
      },
      "CreateCourseRequest": {
        "type": "object",
        "required": [
          "repository"
        ],
        "properties": {
          "repository": {
            "type": "string",
            "description": "The git repository URL of the course"
          }
        }
      },
      "CreateUserCourseRequest": {
        "type": "object",
        "required": [
          "course_slug",
          "proficiency",
          "cadence",
          "accountability"
        ],
        "properties": {
          "accountability": {
            "type": "boolean",
            "description": "Whether the user wants accountability emails"
          },
          "cadence": {
            "type": "string",
            "description": "Practice cadence of the user"
          },
          "course_slug": {
            "type": "string",
            "description": "The slug of the course to enroll in"
          },
          "proficiency": {
            "type": "string",
            "description": "Language proficiency level of the user"
          }
        }
      },
      "ExtensionResponse": {
        "type": "object",
        "required": [
          "slug",
          "name",
          "description",
          "stage_count",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Creation timestamp"
          },
          "description": {
            "type": "string",
            "description": "Extension description"
          },
          "name": {
            "type": "string",
            "description": "Extension name"
          },
          "slug": {
            "type": "string",
            "description": "Unique identifier within course"
          },
          "stage_count": {
            "type": "integer",
            "format": "int32",
            "description": "Number of stages in the extension"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "Last update timestamp"
          }
        }
      },
...
      "UpdateUserCourseRequest": {
        "type": "object",
        "required": [
          "proficiency",
          "cadence",
          "accountability"
        ],
        "properties": {
          "accountability": {
            "type": "boolean",
            "description": "Whether the user wants accountability emails"
          },
          "cadence": {
            "type": "string",
            "description": "Practice cadence of the user"
          },
          "proficiency": {
            "type": "string",
            "description": "Language proficiency level of the user"
          }
        }
      },
      "UserCourseResponse": {
        "type": "object",
        "required": [
          "course_slug",
          "started_at",
          "completed_stage_count",
          "proficiency",
          "cadence",
          "accountability",
          "activated",
          "repository"
        ],
        "properties": {
          "accountability": {
            "type": "boolean",
            "description": "Whether the user wants accountability emails"
          },
          "activated": {
            "type": "boolean",
            "description": "Whether the first Git push was received"
          },
          "cadence": {
            "type": "string",
            "description": "Practice cadence of the user"
          },
          "completed_stage_count": {
            "type": "integer",
            "format": "int32",
            "description": "Number of stages completed by the user"
          },
          "course_slug": {
            "type": "string",
            "description": "Slug of the enrolled course"
          },
          "current_stage_slug": {
            "type": [
              "string",
              "null"
            ],
            "description": "Slug of the current stage the user is on"
          },
          "proficiency": {
            "type": "string",
            "description": "Language proficiency level of the user"
          },
          "repository": {
            "type": "string",
            "description": "The git repository URL of the user course"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp when the enrollment started"
          }
        }
      },
...
    },
    "securitySchemes": {
      "JWTBearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}